@@ -11,15 +11,15 @@ git_source(:github) do |repo_name|
 end
 
 # Bundle edge Rails instead: gem "rails", github: "rails/rails"
-gem "rails", "5.0.2"
+gem "rails", "5.1.5"
 
 # Database
-gem "pg"
+gem "pg", "~> 0.21"
 
 # authentication, authorization, integrations
 gem "attr_encrypted"
 gem "cancancan"
-gem "devise"
+gem "devise", "4.4.1" # 4.4.2 has bug the prevents specs from passing. We can remove this lock when > 4.4.2 is released.
 gem "jwt", "~> 1.5.0" # json web token
 gem "oauth", "~> 0.5.0"
 gem "omniauth"
@@ -72,7 +72,7 @@ end
 group :development, :test do
   gem "byebug", platform: :mri
   gem "dotenv-rails"
-  gem "factory_girl_rails"
+  gem "factory_bot_rails"
   gem "faker"
   gem "guard-rspec", require: false
   gem "rspec-rails"
