@@ -7,7 +7,7 @@ describe ApplicationController, type: :controller do
     respond_to :json
 
     def index
-      render text: "Page: #{@page} Per Page: #{@per_page}"
+      render plain: "Page: #{@page} Per Page: #{@per_page}"
     end
   end
 
