@@ -60,7 +60,7 @@ Rails.application.configure do
                      routing_specs: false,
                      controller_specs: false,
                      request_specs: false
-    g.fixture_replacement :factory_girl, dir: "spec/factories"
+    g.fixture_replacement :factory_bot, dir: "spec/factories"
   end
 
   # Example action_mailer config
